% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgen.R
\name{fitfreq}
\alias{fitfreq}
\title{Allele frequency after one generation of selection}
\usage{
fitfreq(q, h, s)
}
\arguments{
\item{q}{Current frequency of allele A.}

\item{h}{Dominance coefficient of allele A.}

\item{s}{Selection coefficient of allele A.}
}
\value{
The allele frequency of A in the next generation.
}
\description{
Computes the frequency of allele A after one generation of selection at a
single biallelic locus, assuming random mating and a fitness of
\code{1 + h*s} for heterozygotes and \code{1 + s} for AA homozygotes.
}
\examples{
fitfreq(q = 0.1, h = 0.5, s = 0.01)
}
