% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgen.R
\name{ncne}
\alias{ncne}
\title{Nc/Ne ratio from selfing rate, Pollak 1987}
\usage{
ncne(s)
}
\arguments{
\item{s}{Selfing rate in \code{[0, 1]}.}
}
\value{
The ratio of the census size to the effective size.
}
\description{
Nc/Ne ratio from selfing rate, Pollak 1987
}
\examples{
ncne(0.5)
}
