% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{print.expr_matrix}
\alias{print.expr_matrix}
\title{Print an expr_matrix}
\usage{
\method{print}{expr_matrix}(x, ...)
}
\arguments{
\item{x}{An \code{expr_matrix} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The object, invisibly.
}
\description{
Print an expr_matrix
}
\examples{
m <- expr_matrix(matrix(c(10, 0, 5, 8, 2, 7), nrow = 3))
print(m)
}
