% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{print.expr_sample_meta}
\alias{print.expr_sample_meta}
\title{Print an expr_sample_meta object}
\usage{
\method{print}{expr_sample_meta}(x, ...)
}
\arguments{
\item{x}{An \code{expr_sample_meta} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The object, invisibly.
}
\description{
Print an expr_sample_meta object
}
\examples{
d <- expr_sample_meta(data.frame(
  cell_type = c("neuron", "glia"), condition = c("ctrl", "ctrl")
))
print(d)
}
