% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgen.R
\name{print.p0_vec}
\alias{print.p0_vec}
\title{Print a p0_vec object}
\usage{
\method{print}{p0_vec}(x, ...)
}
\arguments{
\item{x}{A \code{p0_vec} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The object, invisibly.
}
\description{
Print a p0_vec object
}
\examples{
p <- p0_vec(c(0.1, 0.5, 0.9))
print(p)
}
