% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgen.R
\name{print.sample_info}
\alias{print.sample_info}
\title{Print a sample_info object}
\usage{
\method{print}{sample_info}(x, ...)
}
\arguments{
\item{x}{A \code{sample_info} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The object, invisibly.
}
\description{
Print a sample_info object
}
\examples{
d <- sample_info(data.frame(
  population = c("AA", "BB"), time_point = c(0, 10), replicate = c("R1", "R1"),
  sample_size = c(30, 30)
))
print(d)
}
