% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgen.R
\name{print.snp_coords}
\alias{print.snp_coords}
\title{Print a snp_coords object}
\usage{
\method{print}{snp_coords}(x, ...)
}
\arguments{
\item{x}{A \code{snp_coords} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The object, invisibly.
}
\description{
Print a snp_coords object
}
\examples{
d <- snp_coords(data.frame(chrom = c("1", "2"), pos = c(100, 200)))
print(d)
}
