% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgen.R
\name{rm_na_after_na}
\alias{rm_na_after_na}
\title{Mark entries after first NA as also NA}
\usage{
rm_na_after_na(x)
}
\arguments{
\item{x}{Numeric vector assumed to be ordered in time.}
}
\value{
The input vector with all entries from the first NA onward set
to NA. Unchanged if there is no NA.
}
\description{
Given a time series of allele frequencies ordered from time 0, 1, 2, ...,
t, sets every entry from the first NA onward to NA, truncating the
trajectory at its first missing value.
}
\examples{
rm_na_after_na(c(0.1, 0.2, NA, 0.4, 0.5))
}
