% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgen.R
\name{arcsin_sqrt}
\alias{arcsin_sqrt}
\title{Arcsine-square root transform of allele frequencies}
\usage{
arcsin_sqrt(x)
}
\arguments{
\item{x}{Numeric vector of values in \code{[0, 1]}; NA entries pass
through as NA.}
}
\value{
The transformed numeric vector, in \code{[0, pi/2]}.
}
\description{
Applies the variance-stabilizing arcsine-square root transformation,
\code{asin(sqrt(x))}, commonly used before temporal analyses of allele
frequencies.
}
\examples{
arcsin_sqrt(c(0, 0.5, 1))
}
