% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgen.R
\name{hill_weir_r2}
\alias{hill_weir_r2}
\title{Expected linkage disequilibrium between loci a given distance apart}
\usage{
hill_weir_r2(d, n, C)
}
\arguments{
\item{d}{Distance between two loci in bp.}

\item{n}{Sample size.}

\item{C}{Population-scaled recombination rate: 4Nc.}
}
\value{
The expected r^2.
}
\description{
Computes the expected r^2 between two loci separated by distance \code{d},
accounting for finite sample size and population-scaled recombination
rate (Hill and Weir 1988).
}
\examples{
hill_weir_r2(d = 1000, n = 30, C = 0.001)
hill_weir_r2(d = c(0, 1000, 10000), n = 30, C = 0.001)
}
