% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgen.R
\name{WF.sel}
\alias{WF.sel}
\title{Deprecated: Wright-Fisher simulation under selection}
\usage{
WF.sel(N, q, h, s, G)
}
\arguments{
\item{N}{Population size (number of diploid individuals).}

\item{q}{Initial frequency of allele A.}

\item{h}{Dominance coefficient of allele A.}

\item{s}{Selection coefficient of allele A. Either a scalar (recycled for
every generation) or a vector of length \code{G - 1} giving the selection
coefficient applied in each generation.}

\item{G}{Number of generations to simulate.}
}
\value{
See \code{\link[=WF_sel]{WF_sel()}}.
}
\description{
\code{\link[=WF.sel]{WF.sel()}} is deprecated; use \code{\link[=WF_sel]{WF_sel()}} instead. Behaves identically.
}
\keyword{internal}
